修改签到接口 新增service：随机生成签到记录

+++ src/main/java/com/ailab/ailabsystem/common/CommonConstant.java
@@ -15,7 +15,7 @@ public class CommonConstant {
     /**
      * 默认签出时间，5小时
      */
-    public static final Integer DEFAULT_CHECK_OUT_TIME = 5;
+    public static final Double DEFAULT_CHECK_OUT_TIME = 5.0;
 
     /**
      * 一个小时
